% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruments_tuning.R
\name{instruments_tuning}
\alias{instruments_tuning}
\alias{instrument}
\alias{mastertune_coarse}
\alias{mastertune_fine}
\title{instruments and tuning}
\usage{
instrument(time, channel = 1, instrument = 1)

mastertune_coarse(time, channel, semitones)

mastertune_fine(time, channel, cents)
}
\arguments{
\item{time}{starting time of note or time of changing instrument in milliseconds}

\item{channel}{channel to use on midi output, in range 1-16,
channel 10 is the drum channel.}

\item{instrument}{instrument number, range 1 to 128}

\item{semitones}{for changing tuning in semitone steps, range -24 to 24}

\item{cents}{for changing tuning in cent steps, range -100 to 100}
}
\description{
Instruments and tuning
}
