% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midi_structures.R
\name{midi_structures}
\alias{midi_structures}
\alias{make_midi_track}
\alias{make_midi_df}
\alias{make_midi_csv}
\alias{make_midi_events}
\title{Midi structures}
\usage{
make_midi_track(..., track = 1, channel = NA)

make_midi_df(..., division = 100)

make_midi_csv(midi_df, midi_csv = NULL)

make_midi_events(...)
}
\arguments{
\item{...}{multiple dataframes containing midi structures}

\item{track}{track number}

\item{channel}{channel number}

\item{division}{for setting playing speed}

\item{midi_df}{midi_df dataframe to be converted to csv}

\item{midi_csv}{csv file to contain midi data to be played}
}
\value{
dataframe containing midi structures
}
\description{
structures representing midi
\itemize{
\item \code{make_midi_df()}: make midi_df object from tracks
}
}
