% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miditoolsR-package.R
\docType{package}
\name{miditoolsR-package}
\alias{miditoolsR}
\alias{miditoolsR-package}
\title{miditoolsR: Tools to create and play MIDI files}
\description{
Midi files (a standard way to represent music as file) are binary files and cannot be handled easily in R. This toolkit allows to create midi files from dataframes by mapping variables to pitches, durations, volume, ans instruments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/neuwirthe/miditoolsR}
  \item Report bugs at \url{https://github.com/neuwirthe/miditoolsR/issues}
}

}
\author{
\strong{Maintainer}: Erich Neuwirth \email{erich.neuwirth@univie.ac.at}

}
\keyword{internal}
