% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_expand_on_off.R
\name{note_manipulation}
\alias{note_manipulation}
\alias{clean_notes_on_off}
\alias{combine_notes_on_off}
\alias{expand_notes_on_off}
\title{note manipulation}
\usage{
clean_notes_on_off(midi_df)

combine_notes_on_off(midi_df)

expand_notes_on_off(midi_df_ext)
}
\arguments{
\item{midi_df}{dataframe of class \code{midi_df}}

\item{midi_df_ext}{dataframe of class \code{midi_df_ext}}
}
\value{
dataframe containing midi structures
}
\description{
Combine \code{Note_on_c} commands and \verb{Note_off_c commands to a single }Note_c\verb{command or expand}Note_c\verb{command to }Note_on_c\code{and}Note_off_c`
commands.
\itemize{
\item \code{clean_notes_off()}: Replace \code{Note_on_c} events with volume 0 by \code{Note_off_c} events
\item \code{combine_notes_on_off()} Replace \code{Note_on_c} and corresponding \code{Note_off_c} events
by \code{Note_c} events.
\item \code{combine_notes_on_off()} Expand \code{Note_c} event to \code{Note_on_c} and corresponding \code{Note_off_c} events.
}
}
