% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\name{notes}
\alias{notes}
\alias{note}
\alias{chord}
\alias{note_on}
\alias{note_off}
\title{Notes}
\usage{
note(time, channel, pitch, velocity, duration)

chord(time, channel, pitch_vec, velocity, duration)

note_on(time, channel, pitch, velocity)

note_off(time, channel, pitch)
}
\arguments{
\item{time}{starting time of note or time of changing instrument in milliseconds}

\item{channel}{channel to use on midi output, in range 1-16,
channel 10 is the drum channel.}

\item{pitch}{range from 0 to 127, pitch (frequency related)}

\item{velocity}{of the note, corresponds to volume of tone}

\item{duration}{duration of sound produced by this command}

\item{pitch_vec}{vector of pitches forming a chord}
}
\value{
tibble containing row(s) defining sound and instrument events
}
\description{
Basic commands to create notes and select instruments
\itemize{
\item \code{note()}: Create notes by combining \code{note_on()} and \code{note_off()}.
\item \code{note_on()}: Create a command which will start playing a note.
\item \code{note_off()}: Create a commandx which will stop playing a note.
\item \code{chord()}: Create command to play notes simultaneously.
}
}
