% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_midi_file.R
\name{read_midi_file}
\alias{read_midi_file}
\title{read midi file}
\usage{
read_midi_file(midi_file)
}
\arguments{
\item{midi_file}{path of a midi file. Needs to be given for input in \code{read_midi_file}.}
}
\value{
midi_df (tibble with MIDI events in tracks)
}
\description{
Read MIDI file into dataframe (tibble)
}
