% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midi_commands_controls.R
\name{sound_modifier_commands}
\alias{sound_modifier_commands}
\alias{pitch_bend}
\alias{pitch_bend_range}
\alias{pan}
\alias{volume}
\alias{expression}
\alias{modulation}
\alias{sustain}
\alias{reverb}
\alias{tremolo}
\alias{chorus}
\title{Midi controls}
\usage{
pitch_bend(time, channel, bend)

pitch_bend_range(time, channel, semitones)

pan(time, channel, position)

volume(time, channel, volume)

expression(time, channel, volume)

modulation(time, channel, on_off)

sustain(time, channel, on_off)

reverb(time, channel, value)

tremolo(time, channel, value)

chorus(time, channel, value)
}
\arguments{
\item{time}{time for applying effect}

\item{channel}{channel number, range 1-16, channel 10 is drum channel}

\item{bend}{amout of pitch bending, range -1 to 1}

\item{semitones}{maximum pitch bend for pitch bend values -1 and 1}

\item{position}{position of instrument in stereo space, range from -1 (left) to 1 (right)}

\item{volume}{volume of sound, range 0-1}

\item{on_off}{switch, 0 is off, 1 is on}

\item{value}{strength of effect, range from 0 to 1}
}
\value{
midi_event dataframe
}
\description{
Commands to modify instrument sounds
}
