% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{tuning}
\alias{tuning}
\alias{mastertune_coarse}
\alias{mastertune_fine}
\title{tuning}
\usage{
mastertune_coarse(time, channel, semitones)

mastertune_fine(time, channel, cents)
}
\arguments{
\item{time}{time of changing tuning in milliseconds}

\item{channel}{channel to use on midi output, in range 1-16}

\item{semitones}{for changing tuning in semitone steps, range -24 to 24}

\item{cents}{for changing tuning in cent steps, range -100 to 100}
}
\description{
Tuning
\itemize{
\item \code{mastertune_coarse()}: changes tuning of instrument in semitones, i.e.
transposing
\item \code{mastertune_fine()}: changes tuning of instruments in cents
(subdivisions of a semitone)
}
}
